# RFC: Reimplement Recycler to reduce impact of pathological cases