Motivation:

Recycler is hard to tune and reduces performance in same cases. Inspired by the
simplicity and large performance gains seen with Go's sync.Pool, which only
reuses objects between GCs, it seemed a more conservative approach could be
applied to Recycler. Note that this looks nothing like sync.Pool, as Recycler
has _very_ different trade-offs and environment.

Modifications:

Reimplement Recycler as a weak reference free list, with low rate of
recycler-enabled objects and heuristics for reusing older objects. Add a
nanobenchmark to better contrast the implementation options, especially with
the 'gc' profiler.

Result:

A Recycler that can improve performance while worst-case would have neglible
impact to GC, memory, and CPU use.

------

There's no rush for this and it wouldn't bother me to close it. I couldn't find the issue where we were talking about disabling the Recycler. This PR should not be considered an argument against doing that; this was a "fun thing" that had interesting enough results I thought I should share. _I honestly don't know if it's worth anyone's time to look into this as a "solution" to Recycler._

I've only run some nanobenchmarks against this. The "generational" behavior in particular is unproven, although it looks interesting. It's possible it is unnecessary and just a single LIFO linked list would be enough, but I went ahead with it because it was the more conservative approach and shouldn't be any worse than the single list approach.

I was a bit worried about whether having so many weak references could cause GC slowdowns; weak refs aren't something I tend to see at large scale. But I think many such concerns are addressed by the 1% sampling.

Although it didn't show up in the code at all, I should note it is _really_ easy to add statistics to the linked list like its length or how reused the objects were, and then have that available _after_ a GC threw the list away by keeping the stats in the WeakRef.

The new implementation doesn't need to use ObjectCleaner :smile:.