@ejona86 2018-06-25
Note that producerConsumer is mainly there for the gc.alloc.rate.norm. The runtimes are coarse, but get the point across; I ran with -Dforks=1 -DwarmupIterations=5 -DmeasureIterations=3

```
before:
RecyclerBenchmark.plainNew                                               avgt    3     9.088 ±    2.504   ns/op
RecyclerBenchmark.plainNew:·gc.alloc.rate                                avgt    3  5597.032 ± 1510.210  MB/sec
RecyclerBenchmark.plainNew:·gc.alloc.rate.norm                           avgt    3    80.000 ±    0.001    B/op
RecyclerBenchmark.plainNew:·gc.churn.PS_Eden_Space                       avgt    3  5601.251 ±  158.963  MB/sec
RecyclerBenchmark.plainNew:·gc.churn.PS_Eden_Space.norm                  avgt    3    80.071 ±   20.797    B/op
RecyclerBenchmark.plainNew:·gc.churn.PS_Survivor_Space                   avgt    3     0.076 ±    0.218  MB/sec
RecyclerBenchmark.plainNew:·gc.churn.PS_Survivor_Space.norm              avgt    3     0.001 ±    0.003    B/op
RecyclerBenchmark.plainNew:·gc.count                                     avgt    3    99.000             counts
RecyclerBenchmark.plainNew:·gc.time                                      avgt    3    68.000                 ms
RecyclerBenchmark.producerConsumer                                       avgt    3   316.584 ±   77.286   ns/op
RecyclerBenchmark.producerConsumer:consumer                              avgt    3   316.583 ±   77.303   ns/op
RecyclerBenchmark.producerConsumer:producer                              avgt    3   316.584 ±   77.271   ns/op
RecyclerBenchmark.producerConsumer:·gc.alloc.rate                        avgt    3   264.830 ±  125.890  MB/sec
RecyclerBenchmark.producerConsumer:·gc.alloc.rate.norm                   avgt    3    65.978 ±   16.503    B/op
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Eden_Space               avgt    3   273.072 ± 1725.475  MB/sec
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Eden_Space.norm          avgt    3    67.844 ±  417.193    B/op
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Survivor_Space           avgt    3     0.021 ±    0.380  MB/sec
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Survivor_Space.norm      avgt    3     0.005 ±    0.094    B/op
RecyclerBenchmark.producerConsumer:·gc.count                             avgt    3     5.000             counts
RecyclerBenchmark.producerConsumer:·gc.time                              avgt    3     5.000                 ms
RecyclerBenchmark.recyclerGetAndOrphan                                   avgt    3    15.122 ±    1.601   ns/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.alloc.rate                    avgt    3  4704.758 ±  520.253  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.alloc.rate.norm               avgt    3   112.000 ±    0.001    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Eden_Space           avgt    3  4754.013 ±   16.992  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Eden_Space.norm      avgt    3   113.175 ±   12.780    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Survivor_Space       avgt    3     0.173 ±    1.160  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Survivor_Space.norm  avgt    3     0.004 ±    0.027    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.count                         avgt    3    84.000             counts
RecyclerBenchmark.recyclerGetAndOrphan:·gc.time                          avgt    3    55.000                 ms
RecyclerBenchmark.recyclerGetAndRecycle                                  avgt    3    14.302 ±    3.369   ns/op
RecyclerBenchmark.recyclerGetAndRecycle:·gc.alloc.rate                   avgt    3    ≈ 10⁻⁴             MB/sec
RecyclerBenchmark.recyclerGetAndRecycle:·gc.alloc.rate.norm              avgt    3    ≈ 10⁻⁵               B/op
RecyclerBenchmark.recyclerGetAndRecycle:·gc.count                        avgt    3       ≈ 0             counts

after (no 1% sampling):
RecyclerBenchmark.plainNew                                               avgt    3     8.983 ±    0.167   ns/op
RecyclerBenchmark.plainNew:·gc.alloc.rate                                avgt    3  5654.780 ±  154.695  MB/sec
RecyclerBenchmark.plainNew:·gc.alloc.rate.norm                           avgt    3    80.000 ±    0.001    B/op
RecyclerBenchmark.plainNew:·gc.churn.PS_Eden_Space                       avgt    3  5658.427 ± 1857.463  MB/sec
RecyclerBenchmark.plainNew:·gc.churn.PS_Eden_Space.norm                  avgt    3    80.050 ±   24.149    B/op
RecyclerBenchmark.plainNew:·gc.churn.PS_Survivor_Space                   avgt    3     0.062 ±    0.001  MB/sec
RecyclerBenchmark.plainNew:·gc.churn.PS_Survivor_Space.norm              avgt    3     0.001 ±    0.001    B/op
RecyclerBenchmark.plainNew:·gc.count                                     avgt    3   100.000             counts
RecyclerBenchmark.plainNew:·gc.time                                      avgt    3    63.000                 ms
RecyclerBenchmark.producerConsumer                                       avgt    3   264.363 ±  952.286   ns/op
RecyclerBenchmark.producerConsumer:consumer                              avgt    3   264.365 ±  952.373   ns/op
RecyclerBenchmark.producerConsumer:producer                              avgt    3   264.360 ±  952.199   ns/op
RecyclerBenchmark.producerConsumer:·gc.alloc.rate                        avgt    3   388.263 ± 1531.383  MB/sec
RecyclerBenchmark.producerConsumer:·gc.alloc.rate.norm                   avgt    3    78.489 ±    9.565    B/op
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Eden_Space               avgt    3   383.914 ± 1746.368  MB/sec
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Eden_Space.norm          avgt    3    77.384 ±   90.381    B/op
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Survivor_Space           avgt    3     0.021 ±    0.380  MB/sec
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Survivor_Space.norm      avgt    3     0.004 ±    0.084    B/op
RecyclerBenchmark.producerConsumer:·gc.count                             avgt    3     7.000             counts
RecyclerBenchmark.producerConsumer:·gc.time                              avgt    3     4.000                 ms
RecyclerBenchmark.recyclerGetAndOrphan                                   avgt    3    23.962 ±    3.783   ns/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.alloc.rate                    avgt    3  3816.896 ±  562.904  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.alloc.rate.norm               avgt    3   144.000 ±    0.001    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Eden_Space           avgt    3  3847.989 ± 1789.199  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Eden_Space.norm      avgt    3   145.180 ±   71.643    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Survivor_Space       avgt    3     0.250 ±    1.317  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Survivor_Space.norm  avgt    3     0.009 ±    0.051    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.count                         avgt    3    68.000             counts
RecyclerBenchmark.recyclerGetAndOrphan:·gc.time                          avgt    3    48.000                 ms
RecyclerBenchmark.recyclerGetAndRecycle                                  avgt    3    18.690 ±    0.170   ns/op
RecyclerBenchmark.recyclerGetAndRecycle:·gc.alloc.rate                   avgt    3    ≈ 10⁻⁴             MB/sec
RecyclerBenchmark.recyclerGetAndRecycle:·gc.alloc.rate.norm              avgt    3    ≈ 10⁻⁵               B/op
RecyclerBenchmark.recyclerGetAndRecycle:·gc.count                        avgt    3       ≈ 0             counts


# Note the runtime difference in recyclerGetAndRecycle and lack of change to recyclerGetAndOrphan
after (no 1% sampling, non-generational):
RecyclerBenchmark.plainNew                                               avgt    3     9.029 ±    1.148   ns/op
RecyclerBenchmark.plainNew:·gc.alloc.rate                                avgt    3  5629.562 ±  710.186  MB/sec
RecyclerBenchmark.plainNew:·gc.alloc.rate.norm                           avgt    3    80.000 ±    0.001    B/op
RecyclerBenchmark.plainNew:·gc.churn.PS_Eden_Space                       avgt    3  5605.200 ± 3104.981  MB/sec
RecyclerBenchmark.plainNew:·gc.churn.PS_Eden_Space.norm                  avgt    3    79.647 ±   35.767    B/op
RecyclerBenchmark.plainNew:·gc.churn.PS_Survivor_Space                   avgt    3     0.076 ±    0.580  MB/sec
RecyclerBenchmark.plainNew:·gc.churn.PS_Survivor_Space.norm              avgt    3     0.001 ±    0.008    B/op
RecyclerBenchmark.plainNew:·gc.count                                     avgt    3    99.000             counts
RecyclerBenchmark.plainNew:·gc.time                                      avgt    3    63.000                 ms
RecyclerBenchmark.producerConsumer                                       avgt    3   237.773 ± 1211.562   ns/op
RecyclerBenchmark.producerConsumer:consumer                              avgt    3   237.771 ± 1211.542   ns/op
RecyclerBenchmark.producerConsumer:producer                              avgt    3   237.775 ± 1211.582   ns/op
RecyclerBenchmark.producerConsumer:·gc.alloc.rate                        avgt    3   466.422 ± 2551.443  MB/sec
RecyclerBenchmark.producerConsumer:·gc.alloc.rate.norm                   avgt    3    82.588 ±   24.188    B/op
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Eden_Space               avgt    3   439.561 ± 1746.286  MB/sec
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Eden_Space.norm          avgt    3    81.939 ±  546.164    B/op
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Survivor_Space           avgt    3     0.014 ±    0.219  MB/sec
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Survivor_Space.norm      avgt    3     0.003 ±    0.045    B/op
RecyclerBenchmark.producerConsumer:·gc.count                             avgt    3     8.000             counts
RecyclerBenchmark.producerConsumer:·gc.time                              avgt    3     8.000                 ms
RecyclerBenchmark.recyclerGetAndOrphan                                   avgt    3    23.661 ±    4.677   ns/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.alloc.rate                    avgt    3  3864.795 ±  812.263  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.alloc.rate.norm               avgt    3   144.000 ±    0.001    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Eden_Space           avgt    3  3903.223 ±   47.398  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Eden_Space.norm      avgt    3   145.445 ±   30.515    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Survivor_Space       avgt    3     0.277 ±    1.796  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Survivor_Space.norm  avgt    3     0.010 ±    0.066    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.count                         avgt    3    69.000             counts
RecyclerBenchmark.recyclerGetAndOrphan:·gc.time                          avgt    3    43.000                 ms
RecyclerBenchmark.recyclerGetAndRecycle                                  avgt    3    13.730 ±    0.977   ns/op
RecyclerBenchmark.recyclerGetAndRecycle:·gc.alloc.rate                   avgt    3    ≈ 10⁻⁴             MB/sec
RecyclerBenchmark.recyclerGetAndRecycle:·gc.alloc.rate.norm              avgt    3    ≈ 10⁻⁵               B/op
RecyclerBenchmark.recyclerGetAndRecycle:·gc.count                        avgt    3       ≈ 0             counts


after (1% sampling; this is the code you actually see):
RecyclerBenchmark.plainNew                                               avgt    3     9.053 ±    2.028   ns/op
RecyclerBenchmark.plainNew:·gc.alloc.rate                                avgt    3  5616.314 ± 1301.369  MB/sec
RecyclerBenchmark.plainNew:·gc.alloc.rate.norm                           avgt    3    80.000 ±    0.001    B/op
RecyclerBenchmark.plainNew:·gc.churn.PS_Eden_Space                       avgt    3  5601.811 ± 3141.269  MB/sec
RecyclerBenchmark.plainNew:·gc.churn.PS_Eden_Space.norm                  avgt    3    79.781 ±   26.312    B/op
RecyclerBenchmark.plainNew:·gc.churn.PS_Survivor_Space                   avgt    3     0.090 ±    0.791  MB/sec
RecyclerBenchmark.plainNew:·gc.churn.PS_Survivor_Space.norm              avgt    3     0.001 ±    0.011    B/op
RecyclerBenchmark.plainNew:·gc.count                                     avgt    3    99.000             counts
RecyclerBenchmark.plainNew:·gc.time                                      avgt    3    67.000                 ms
RecyclerBenchmark.producerConsumer                                       avgt    3   194.900 ±  351.183   ns/op
RecyclerBenchmark.producerConsumer:consumer                              avgt    3   194.902 ±  351.094   ns/op
RecyclerBenchmark.producerConsumer:producer                              avgt    3   194.897 ±  351.273   ns/op
RecyclerBenchmark.producerConsumer:·gc.alloc.rate                        avgt    3   296.171 ±  552.715  MB/sec
RecyclerBenchmark.producerConsumer:·gc.alloc.rate.norm                   avgt    3    45.155 ±    4.967    B/op
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Eden_Space               avgt    3   273.668 ± 1741.189  MB/sec
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Eden_Space.norm          avgt    3    41.460 ±  244.824    B/op
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Survivor_Space           avgt    3     0.035 ±    0.580  MB/sec
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Survivor_Space.norm      avgt    3     0.005 ±    0.092    B/op
RecyclerBenchmark.producerConsumer:·gc.count                             avgt    3     5.000             counts
RecyclerBenchmark.producerConsumer:·gc.time                              avgt    3     3.000                 ms
RecyclerBenchmark.recyclerGetAndOrphan                                   avgt    3    12.599 ±    6.015   ns/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.alloc.rate                    avgt    3  4083.809 ± 1873.270  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.alloc.rate.norm               avgt    3    81.000 ±    0.001    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Eden_Space           avgt    3  4072.825 ± 3085.454  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Eden_Space.norm      avgt    3    80.764 ±   30.359    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Survivor_Space       avgt    3     0.062 ±    0.378  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Survivor_Space.norm  avgt    3     0.001 ±    0.007    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.count                         avgt    3    72.000             counts
RecyclerBenchmark.recyclerGetAndOrphan:·gc.time                          avgt    3    50.000                 ms
RecyclerBenchmark.recyclerGetAndRecycle                                  avgt    3    19.069 ±    4.756   ns/op
RecyclerBenchmark.recyclerGetAndRecycle:·gc.alloc.rate                   avgt    3    ≈ 10⁻⁴             MB/sec
RecyclerBenchmark.recyclerGetAndRecycle:·gc.alloc.rate.norm              avgt    3    ≈ 10⁻⁵               B/op
RecyclerBenchmark.recyclerGetAndRecycle:·gc.count                        avgt    3       ≈ 0             counts
```